pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "spredd-api"
version = "0.1.0"
description = "Spredd Developer API — Prediction market trading API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.27.0",
    "web3>=7.0.0",
    "eth-account>=0.13.0",
    "solders>=0.21.0",
    "solana>=0.35.0",
    "opinion-clob-sdk>=0.1.0",
    "limitless-sdk>=0.1.0",
    "python-dotenv>=1.0.0",
    "python-dateutil>=2.9.0",
    "base58>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]
